More compatible zone handling
Overlapping zones now work.  I didn't realize Sphere supported this,
although it makes sense in hindsight.  Person location and speed are now
double precision, as well.

+++ minisphere/map_engine.h
@@ -12,7 +12,7 @@ extern const obsmap_t*  get_map_layer_obsmap    (int layer);
 extern point3_t         get_map_origin          (void);
 extern int              get_map_tile            (int x, int y, int layer);
 extern const tileset_t* get_map_tileset         (void);
-extern void             normalize_map_entity_xy (float* inout_x, float* inout_y, int layer);
+extern void             normalize_map_entity_xy (double* inout_x, double* inout_y, int layer);
 
 extern bool g_map_running;
